+++ database/upgrade/upgrade5.7.0RC.sql
@@ -9,7 +9,7 @@
 
 START TRANSACTION;
 
-DROP TABLE IF EXISTS `myems_system_db`.`tbl_energy_storage_containers_sensors` ;
+DROP TABLE IF EXISTS `myems_system_db`.`tbl_energy_storage_containers_sensors`;
 
 CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_energy_storage_containers_data_sources` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
@@ -26,6 +26,14 @@ CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_microgrids_data_sources` (
   PRIMARY KEY (`id`));
 CREATE INDEX `tbl_microgrids_data_sources_index_1`
 ON `myems_system_db`.`tbl_microgrids_data_sources` (`microgrid_id`);
+
+CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_photovoltaic_power_stations_data_sources` (
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `photovoltaic_power_station_id` BIGINT NOT NULL,
+  `data_source_id` BIGINT NOT NULL,
+  PRIMARY KEY (`id`));
+CREATE INDEX `tbl_photovoltaic_power_stations_data_sources_index_1`
+ON `myems_system_db`.`tbl_photovoltaic_power_stations_data_sources` (`photovoltaic_power_station_id`);
 
 -- UPDATE VERSION NUMBER
 UPDATE `myems_system_db`.`tbl_versions` SET version='5.7.0RC', release_date='2025-07-21' WHERE id=1;