+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added quick mode for svg in myems-admin and myems-api
 - added myems_system_db.tbl_energy_storage_containers_data_sources in database
 - added myems_system_db.tbl_microgrids_data_sources in database
+- added myems_system_db.tbl_photovoltaic_power_stations_data_sources in database
 ### Changed
 - subspace names under the same parent space cannot be duplicated
 - updated power stations